Revert "refactor(frontend): move editor string provider to frontend client base"
This reverts commit b7fd665860291f78ac371fafe9995de693db2c92.
