+++ apps/web/src/components/frontend-client-base/frontend-client-base.tsx
@@ -1,14 +1,7 @@
-import { EditStringsProvider } from '@editor/i18n/edit-strings-provider'
-import { StaticStringsProvider } from '@editor/i18n/static-strings-provider'
-import { editStrings as editStringsDe } from '@editor/i18n/strings/de/edit'
-import { staticStrings as staticStringsDe } from '@editor/i18n/strings/de/static'
-import { editStrings as editStringsEn } from '@editor/i18n/strings/en/edit'
-import { staticStrings as staticStringsEn } from '@editor/i18n/strings/en/static'
 import type { AuthorizationPayload } from '@serlo/authorization'
 import Head from 'next/head'
 import { Router, useRouter } from 'next/router'
 import NProgress from 'nprogress'
-import { mergeDeepRight } from 'ramda'
 import { useState, useEffect } from 'react'
 import { Toaster } from 'react-hot-toast'
 
@@ -120,54 +113,38 @@ export function FrontendClientBase({
   //console.dir(initialProps)
 
   return (
-    <StaticStringsProvider
-      value={
-        instanceData.lang === 'de'
-          ? mergeDeepRight(staticStringsEn, staticStringsDe)
-          : staticStringsEn
-      }
-    >
-      <EditStringsProvider
-        value={
-          instanceData.lang === 'de'
-            ? mergeDeepRight(editStringsEn, editStringsDe)
-            : editStringsEn
-        }
-      >
-        <InstanceDataProvider value={instanceData}>
-          <PrintMode />
-          {noIndex ? (
-            <Head>
-              <meta name="robots" content="noindex" />
-            </Head>
-          ) : null}
-          <AuthProvider unauthenticatedAuthorizationPayload={authorization}>
-            <LoggedInDataProvider value={loggedInData}>
-              <UuidsProvider value={serloEntityData ?? null}>
-                <Toaster />
-                <ConditionalWrap
-                  condition={!noHeaderFooter}
-                  wrapper={(kids) => <HeaderFooter>{kids}</HeaderFooter>}
-                >
-                  <ConditionalWrap
-                    condition={!noContainers}
-                    wrapper={(kids) => (
-                      <div className="relative">
-                        <MaxWidthDiv showNav={showNav}>
-                          <main id="content">{kids}</main>
-                        </MaxWidthDiv>
-                      </div>
-                    )}
-                  >
-                    {children}
-                  </ConditionalWrap>
-                  <MaintenanceBanner />
-                </ConditionalWrap>
-              </UuidsProvider>
-            </LoggedInDataProvider>
-          </AuthProvider>
-        </InstanceDataProvider>
-      </EditStringsProvider>
-    </StaticStringsProvider>
+    <InstanceDataProvider value={instanceData}>
+      <PrintMode />
+      {noIndex ? (
+        <Head>
+          <meta name="robots" content="noindex" />
+        </Head>
+      ) : null}
+      <AuthProvider unauthenticatedAuthorizationPayload={authorization}>
+        <LoggedInDataProvider value={loggedInData}>
+          <UuidsProvider value={serloEntityData ?? null}>
+            <Toaster />
+            <ConditionalWrap
+              condition={!noHeaderFooter}
+              wrapper={(kids) => <HeaderFooter>{kids}</HeaderFooter>}
+            >
+              <ConditionalWrap
+                condition={!noContainers}
+                wrapper={(kids) => (
+                  <div className="relative">
+                    <MaxWidthDiv showNav={showNav}>
+                      <main id="content">{kids}</main>
+                    </MaxWidthDiv>
+                  </div>
+                )}
+              >
+                {children}
+              </ConditionalWrap>
+              <MaintenanceBanner />
+            </ConditionalWrap>
+          </UuidsProvider>
+        </LoggedInDataProvider>
+      </AuthProvider>
+    </InstanceDataProvider>
   )
 }
